+++ app/src/main/java/woowacourse/paint/view/canvas/CanvasView.kt
@@ -5,7 +5,6 @@ import android.graphics.Canvas
 import android.util.AttributeSet
 import android.view.MotionEvent
 import android.view.View
-import woowacourse.paint.view.canvas.paintHistory.PainterHistory
 import woowacourse.paint.view.palette.PaletteMode
 import woowacourse.paint.view.palette.color.PaletteColor
 import woowacourse.paint.view.palette.shape.PaletteShape
