+++ app/src/main/java/woowacourse/paint/view/canvas/painter/Painter.kt
@@ -2,7 +2,7 @@ package woowacourse.paint.view.canvas.painter
 
 import android.graphics.Canvas
 import android.graphics.Paint
-import woowacourse.paint.view.canvas.paintHistory.PainterState
+import woowacourse.paint.view.canvas.PainterState
 import woowacourse.paint.view.palette.PaletteMode
 import woowacourse.paint.view.palette.color.PaletteColor
 import woowacourse.paint.view.palette.shape.PaletteShape
